initial infrastructure's dal commit

+++ core/src/main/java/com/laurenci/hydrotrack/core/record/RecordRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.laurenci.hydrotrack.core.Repository;
 
 public interface RecordRepository extends Repository<Record, RecordId> {
-    List<Record> findByUserId(Integer userId);
+    List<Record> findByUserId(Long userId);
 }
