+++ core/src/main/java/com/laurenci/hydrotrack/core/report/ReportRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.laurenci.hydrotrack.core.Repository;
 
 public interface ReportRepository extends Repository<Report, ReportId> {
-    List<Report> findByUserId(Integer userId);
+    List<Report> findByUserId(Long userId);
 }
